working from the smaller socket server version
